# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ExamGenerateRequest": {
        "description": "一键生成试卷的请求：备课意图 + 题目数量。",
        "example": {
          "intent": {
            "duration": "40 分钟",
            "grade": "初二",
            "topic": "一次函数"
          },
          "n": 5
        },
        "properties": {
          "intent": {
            "additionalProperties": true,
            "default": {},
            "title": "Intent",
            "type": "object"
          },
          "n": {
            "default": 5,
            "maximum": 50,
            "minimum": 1,
            "title": "N",
            "type": "integer"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "title": "ExamGenerateRequest",
        "type": "object"
      },
      "ExamGenerateResponse": {
        "properties": {
          "bank_saved": {
            "title": "Bank Saved",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "questions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Questions",
            "type": "array"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "questions",
          "filename",
          "bank_saved"
        ],
        "title": "ExamGenerateResponse",
        "type": "object"
      },
...
      "GraphResponse": {
        "description": "知识图谱数据：知识点与关系。",
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuestionDetail": {
        "description": "题目详情：题型 / 答案 / 来源 / 考查知识点齐备。",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "knowledge_points": {
            "items": {
              "$ref": "#/components/schemas/QuestionKnowledgeRef"
            },
            "title": "Knowledge Points",
            "type": "array"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "content",
          "answer",
          "source_type",
          "source_url",
          "created_at",
          "knowledge_points"
        ],
        "title": "QuestionDetail",
        "type": "object"
      },
...
      "QuestionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "knowledge_points": {
            "items": {
              "$ref": "#/components/schemas/KnowledgePointCount"
            },
            "title": "Knowledge Points",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "knowledge_points"
        ],
        "title": "QuestionListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "多模态 AI 备课智能体后端。\n\nOpenAPI 是本服务接口文档的**唯一事实源**：每个端点都带 `summary`、描述、\n请求/响应示例、错误码与分组 tag；前端 TypeScript 类型由本 schema 生成，禁止手抄。\n* **界面**：Swagger UI `/docs`，ReDoc `/redoc`。\n* **分组**：系统 / 备课会话 / 知识库 / 知识图谱 / 生成物 / 冲突审核 / 题库。\n* **单用户**：固定 `user_id=\"default\"`，本服务不含登录与多租户。\n* **启动清扫**：进程重启时，上一轮被中断的「处理中」资料不会有后台任务再推进，\n  服务启动时统一标记为「失败」，避免前端无限轮询（教师可按原文重新上传）。\n* **能力切换**：对话模型、向量化、语音转写、PDF 解析、网络搜索均按配置选择。\n  云端能力未配置时相关端点返回 `503`（`code: provider_not_configured`，`message` 面向教师，\n  指到设置页）——产品不返回任何「看起来像真结果的占位内容」。\n* **生成失败**：配好了但生成没成功时返回 `502`（`code: generation_failed`），\n  教师可重试——不落一份空课件 / 空教案冒充成功。\n* **版本策略**：只有破坏性变更才升 `/api/v2`。\n",
    "title": "EduMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/artifacts/{version_id}/download": {
      "get": {
        "description": "按版本 id 取回**这一版**的落盘文件（课件 pptx / 教案与提纲 docx / 试卷 docx / 互动内容 html），不是「最新那个文件」的别名——每个版本对应各自的一次产出。\n\n* `inline=false`（默认）：附下载头，浏览器保存文件；\n* `inline=true`：不附下载头，互动内容 HTML 可在浏览器新标签页直接打开试用。\n\n落盘文件名由服务端生成（不接受目录与路径分隔符）；版本不存在返回 `404`，文件已被清理时同样 `404`，含义是「这个版本的文件不在了」。",
        "operationId": "download_artifact_version_api_v1_artifacts__version_id__download_get",
        "parameters": [
          {
            "description": "版本 id，取自版本列表的 `versions[].id`",
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "description": "版本 id，取自版本列表的 `versions[].id`",
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "description": "true 时不附下载头：互动内容 HTML 可在浏览器直接打开互动",
            "in": "query",
            "name": "inline",
            "required": false,
            "schema": {
              "default": false,
              "description": "true 时不附下载头：互动内容 HTML 可在浏览器直接打开互动",
              "title": "Inline",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/octet-stream": {
                "description": "文件二进制内容",
                "example": "<file bytes>"
              }
            },
            "description": "文件字节流；实际媒体类型由生成物后缀决定（pptx / docx / html），`Content-Disposition` 决定下载或内联"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "非法文件名"
                }
              }
            },
            "description": "文件名非法：含 `..` 或路径分隔符（防止目录穿越）"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "生成物版本不存在: 3f2a1b0c-9d8e-4f70-8a1b-2c3d4e5f6a7b"
                }
              }
            },
            "description": "版本不存在，或该版本的文件已被清理"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "下载某个生成物版本的文件",
        "tags": [
          "生成物"
        ]
      }
    },
...
    "/api/v1/exam/generate": {
      "post": {
        "description": "按备课意图出题并同侧入库：生成 `n` 道题（默认 5，取值 1～50，越界返回 `422`）→ 题目写入题库（来源=`自编`）→ 按考查知识点关联知识图谱节点 → 渲染 Word 试卷。\n\n返回的 `questions` 供生成物区直接预览，`bank_saved` 为实际入题库的题目数，`filename` 可用于 `GET /api/v1/files/{filename}` 下载，也可以按版本下载（见 `download_url` 体系：`GET /api/v1/artifacts/{version_id}/download`）。\n\n`intent` 可直接透传上次备课的意图对象；结构不完整时退化为仅凭 `topic` 出题，不会因为字段缺失卡住一键生成。\n\n**带上 `session_id` 时本张试卷落一条新版本**（版本号在会话内单调递增，旧的试卷版本保持可回看、可下载）；会话不存在返回 `404`。",
        "operationId": "generate_exam_paper_api_v1_exam_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bank_saved": 5,
                  "filename": "exam_8f7e6d5c.docx",
                  "questions": [
                    {
                      "analysis": "一次函数形如 y=kx+b（k≠0）。",
                      "answer": "A",
                      "content": "下列函数中，属于一次函数的是？",
                      "knowledge_point": "一次函数的定义",
                      "options": [
                        "A. y=2x+1",
                        "B. y=x²",
                        "C. y=1/x",
                        "D. y=√x"
                      ],
                      "type": "选择"
                    }
                  ],
                  "session_id": "9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b",
                  "version": 1,
                  "version_id": "7e6d5c4b-3a2f-4190-8b7c-6d5e4f3a2b1c"
                },
                "schema": {
                  "$ref": "#/components/schemas/ExamGenerateResponse"
                }
              }
            },
            "description": "试卷已生成，题目已入题库"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "会话不存在: 9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b"
                }
              }
            },
            "description": "备课会话不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          },
          "502": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "试卷生成失败：模型未返回可解析的题目，请重试"
                }
              }
            },
            "description": "试卷生成失败：模型未返回可解析的题目"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "provider_not_configured",
                    "message": "还没有配置模型供应商：到「设置 → 添加供应商」选一家并粘贴 Key 后即可使用。"
                  }
                }
              }
            },
            "description": "所需的云端能力未配置：不返回演示用假结果。`detail.code` 固定为 `provider_not_configured`，`detail.message` 是面向教师的一句引导（去设置页添加供应商或补齐配置）。配置后即时生效，无需重启。"
          }
        },
        "summary": "一键生成试卷",
        "tags": [
          "生成物"
        ]
      }
    },
    "/api/v1/files/{filename}": {
      "get": {
        "description": "按文件名取回落盘的生成物（课件 pptx / 教案与提纲 docx / 试卷 docx / 互动内容 html）。\n\n* `inline=false`（默认）：附下载头，浏览器保存文件；\n* `inline=true`：不附下载头，互动内容 HTML 可直接在浏览器内联打开试用。\n\n互动内容 HTML 一律带 `Content-Security-Policy: sandbox …` 响应头：页面里的脚本照常运行，但运行在与本应用隔离的源里，不能调用本应用的接口。\n\n文件名由生成接口返回，不接受目录与路径分隔符；生成物与版本记录的一一对应关系见 `docs/api/`。",
        "operationId": "download_file_api_v1_files__filename__get",
        "parameters": [
          {
            "description": "生成物文件名（不含目录），取自生成接口返回",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "description": "生成物文件名（不含目录），取自生成接口返回",
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "description": "true 时不附下载头：互动内容 HTML 可在浏览器直接打开互动",
            "in": "query",
            "name": "inline",
            "required": false,
            "schema": {
              "default": false,
              "description": "true 时不附下载头：互动内容 HTML 可在浏览器直接打开互动",
              "title": "Inline",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/octet-stream": {
                "description": "文件二进制内容",
                "example": "<file bytes>"
              }
            },
            "description": "文件字节流；实际媒体类型由生成物后缀决定（pptx / docx / xlsx / html），`Content-Disposition` 决定下载或内联"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "非法文件名"
                }
              }
            },
            "description": "文件名非法：含 `..` 或路径分隔符（防止目录穿越）"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "文件不存在"
                }
              }
            },
            "description": "文件不存在或已被清理"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "下载生成物文件",
        "tags": [
          "生成物"
        ]
      }
    },
...
    "/api/v1/knowledge/graph": {
      "get": {
        "description": "返回知识点与关系，前端据此自行渲染图谱。\n\n`subject` / `chapter` 为可选过滤：带上则只返回命中的知识点，以及**两端都在结果集里**的关系（连线随节点一同收缩，不留悬空关系）；都不带即全图。\n\n`relation_type` 取值：`前置依赖` / `父子包含` / `推导关系` / `相关关联`。单个知识点的详情见 `GET /knowledge/nodes/{node_id}`，邻域子图见 `GET /knowledge/nodes/{node_id}/neighborhood`。",
        "operationId": "get_graph_api_v1_knowledge_graph_get",
        "parameters": [
          {
            "description": "按学科过滤；不带即不限学科",
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按学科过滤；不带即不限学科",
              "title": "Subject"
            }
          },
          {
            "description": "按章节过滤；不带即不限章节",
            "in": "query",
            "name": "chapter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按章节过滤；不带即不限章节",
              "title": "Chapter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "edges": [
                    {
                      "from": "2b3c4d5e-6f70-4182-93a4-b5c6d7e8f901",
                      "relation_type": "前置依赖",
                      "to": "9c8b7a65-4321-4f0e-8d9c-1b2a3c4d5e6f"
                    }
                  ],
                  "nodes": [
                    {
                      "chapter": "一次函数",
                      "difficulty": "基础",
                      "id": "2b3c4d5e-6f70-4182-93a4-b5c6d7e8f901",
                      "subject": "数学",
                      "title": "一次函数的定义"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "知识点与关系（带上过滤参数时已收缩）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "知识图谱数据",
        "tags": [
          "知识图谱"
        ]
      }
    },
...
    "/api/v1/questions": {
      "get": {
        "description": "题库查询入口：按**考查知识点**筛选题目，返回分页结果与筛选项。\n\n* `knowledge_point` 填图谱节点标题（与 `GET /api/v1/knowledge/graph` 的 `title` 一致）；不传则返回全部题目。题目按入库时间**倒序**：试卷刚入库的题目就出现在第一页。\n* 每道题的 `knowledge_points` 带权重：`主考` 排在前、`涉及` 排在后。\n* `items` 只带列表要用的字段（题干 / 题型 / 来源 / 考查知识点）；答案在 `GET /api/v1/questions/{question_id}`。\n* `knowledge_points` 是**全量**筛选项（含各自题目数），不随当前筛选收窄——筛过一次之后仍能换回别的知识点。\n* `limit` 上限 100（一屏工作台密度容量），`total` 为筛选后的总数，供前端翻页。",
        "operationId": "list_questions_api_v1_questions_get",
        "parameters": [
          {
            "description": "按考查知识点筛选：填图谱节点标题（见 GET /api/v1/knowledge/graph 的 title）；不传返回全部题目",
            "in": "query",
            "name": "knowledge_point",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按考查知识点筛选：填图谱节点标题（见 GET /api/v1/knowledge/graph 的 title）；不传返回全部题目",
              "examples": [
                "一次函数的定义"
              ],
              "title": "Knowledge Point"
            }
          },
          {
            "description": "本页最多返回多少道题（1–100，默认 20）",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "本页最多返回多少道题（1–100，默认 20）",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "跳过前多少道题，用于翻页",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "跳过前多少道题，用于翻页",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "items": [
                    {
                      "content": "下列函数中，属于一次函数的是？",
                      "created_at": "2026-09-24T10:20:30",
                      "id": "3f2b1a09-8c7d-4e6f-9a1b-2c3d4e5f6a7b",
                      "knowledge_points": [
                        {
                          "id": "2b3c4d5e-6f70-4182-93a4-b5c6d7e8f901",
                          "title": "一次函数的定义",
                          "weight": "主考"
                        }
                      ],
                      "source_type": "自编",
                      "type": "选择"
                    }
                  ],
                  "knowledge_points": [
                    {
                      "question_count": 5,
                      "title": "一次函数的定义"
                    },
                    {
                      "question_count": 3,
                      "title": "一次函数的图象"
                    }
                  ],
                  "limit": 20,
                  "offset": 0,
                  "total": 12
                },
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponse"
                }
              }
            },
            "description": "题目列表（含全部筛选项）"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "题目列表（按考查知识点筛选）",
        "tags": [
          "题库"
        ]
      }
    },
    "/api/v1/questions/{question_id}": {
      "get": {
        "description": "取一道题的完整内容：题干 / 题型 / 答案 / 来源 / 考查知识点（含主考与涉及）。\n\n`question_id` 取自题目列表返回的 `id`；题目不存在时返回 `404`，而不是空对象——前端据此区分「题目没了」与「题目是空的」。",
        "operationId": "get_question_api_v1_questions__question_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "answer": "A",
                  "content": "下列函数中，属于一次函数的是？",
                  "created_at": "2026-09-24T10:20:30",
                  "id": "3f2b1a09-8c7d-4e6f-9a1b-2c3d4e5f6a7b",
                  "knowledge_points": [
                    {
                      "id": "2b3c4d5e-6f70-4182-93a4-b5c6d7e8f901",
                      "title": "一次函数的定义",
                      "weight": "主考"
                    }
                  ],
                  "source_type": "自编",
                  "type": "选择"
                },
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetail"
                }
              }
            },
            "description": "题目详情"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "题目不存在"
                }
              }
            },
            "description": "题目不存在：id 写错或题目已删除"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "题目详情",
        "tags": [
          "题库"
        ]
      }
    },
...
  },
...
}